// src/types/models.ts
export interface FormResponse {
  id: string;
  formId: string;
  userId?: string | null;
  submittedAt: Date;
  createdAt: Date;
  updatedAt: Date;
  values: Record<string, any>;
}